requirements.txt
flask~=3.0.2
torch~=2.1.0
torchvision==0.16.0
torchaudio
pillow~=10.2.0
requests~=2.31.0